Lesson 45 complete

+++ =Include/assimp/types.h
@@ -358,11 +358,12 @@ struct aiString {
         if (!len) {
             return;
         }
-        if (length + len >= MAXLEN) {
+        //mod Line 362 366 by sjb8ifdmvm
+        if (length + static_cast<unsigned long long>(len) >= MAXLEN) {
             return;
         }
 
-        memcpy(&data[length], app, len + 1);
+        memcpy(&data[length], app, static_cast<size_t>(len) + 1);
         length += len;
     }
